Extract numeric operator helpers
